% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd2utm.R
\name{dd2utm}
\alias{dd2utm}
\title{Decimal degrees to Universal Transverse Mercator}
\usage{
dd2utm(id, long, lat, data)
}
\arguments{
\item{id}{The column number containing the identifier for each row. This 'identifier' may be the name of a population/provenance or could be a sequence of numbers}

\item{long}{The column number containing the logitude decimal degree coordinates}

\item{lat}{The column number containing the latitude decimal degree coordinates}

\item{data}{The data frame containing the latitude/longitude coodinates. This can be of either class 'data.frame' or 'matrix'}
}
\value{
An object of class 'data.frame' containing four columns: 
\describe{
\item{id}{The identifier from the original data file}
\item{north}{The north UTM coordinate}
\item{east}{The east UTM coordinate}
\item{zone}{The zone which the latitudes originate. Zone ranges from 1 (180 to 147 W) to 60}}
}
\description{
A function to convert decimal degrees into universal transverse mercator coordinate system. This is an R implementation of the formulas provided by Prof. Steven Dutch.
}
\examples{
# Create a data frame
tmp <- data.frame(id = "test", longitude = 147.4686, latitude = -42.008)

# Run function
PUCA::dd2utm(id = 1, long = 2, lat = 3, data = tmp)

# Output
#     id   north     east zone
# 1 test 5349229 538803.5   55

}
\author{
P.A. Harrison
}
\references{
CFF Karney (2010) Transverse Mercator with an accuracy of a few nanometers, http://arxiv.org/abs/1002.1417v3

\url{http://www.uwgb.edu/dutchs/UsefulData/UTMFormulas.HTM}
}

