% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawBuffer.R
\name{drawBuffer}
\alias{drawBuffer}
\title{Draw a buffer around a point}
\usage{
drawBuffer(x, y, r, numVert = 1000, n = 20, mask = NULL)
}
\arguments{
\item{x}{The longitude coordinate for the centre of the buffer.}

\item{y}{The latitude coordinate for the centre of the buffer.}

\item{r}{The radius (in kilometers) of the buffer.}

\item{numVert}{The number of vertices to create the buffer.}

\item{n}{The number of uniform points to create within the buffer}

\item{mask}{Optinal mask shapefile (.shp) used to clip the buffer.}
}
\value{
An object of class 'SpatialPoints'
}
\description{
A helper function for function \code{seedSource}. This function first creates a a polygon of class 'SpatialPolygonsDataFrame' and subsamples the polygon to create a uniform point patten of class 'SpatialPoints' which is retunred to an object.
}
\examples{
#' #load libraries
library(raster)
library(sp)

tmp <- PUCA::drawBuffer(x = 147.4686, y = -42.008, r = 5)
plot(tmp)

}
\author{
P. A. Harrison
}
\seealso{
\code{\link{seedSource}}, \code{\link[plotrix]{draw.circle}}, \code{\link[sp]{SpatialPolygons}}, \code{\link[sp]{SpatialPolygonsDataFrame}}, \code{\link[sp]{spsample}}
}

