% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{geoCode}
\alias{geoCode}
\title{Georeference a set of coordinates}
\usage{
geoCode(address, print = F, type = 1)
}
\arguments{
\item{address}{The latitude and longitude coordinates for a point of interest seperated by a comma, as class 'character'}

\item{print}{Logical argument whether to print the results to the consol}

\item{type}{The type of address to return. See Details for arguments.}
}
\value{
A list of georeferenced locations
}
\description{
geoCode is a helper function for \code{georefPoints} and quries the Google Map API server to return the street address for a set of coordinates.
}
\details{
The function \code{geoCode} returns various annotations for a set of coordinates:

type = 1 : returns the full street address

type = 2 : returns the suburb with no street identifier

type = 3 : returns the state and postcode (zipcode)

type = 4 : return the state and country
}
\section{Warning}{

The Google Map API has a maximum threshold limit on the number of points that can be georeferenced at any one time. Preliminary testing has found >1000 points in one request can result in the code stopping and returning an error message. There is also a threshold on the number of requests to the API server in a single day, and may result in the code stopping with an error message. If this occurs the threshold is reset in 24 hours. See \url{https://developers.google.com/maps/} for more details.
}
\examples{
# Load libraries
library(RJSONIO)

# Create data frame
tmp <- data.frame(id = "test", longitude = 147.4686, latitude = -42.008)


}
\author{
P. A. Harrison
}
\seealso{
\code{\link{georefPoints}}
}

