% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefPoints.R
\name{georefPoints}
\alias{georefPoints}
\title{Georeference a set of coordinates from \code{seedSource}}
\usage{
georefPoints(data, rank, n, printToScreen = T)
}
\arguments{
\item{data}{The output data frame of \code{seedSource}}

\item{rank}{The column name used to rank the data frame}

\item{n}{The number of points to georeference}

\item{printToScreen}{Logic whether to print results to screen. Default is \code{TRUE}}
}
\value{
A data frame showing the coordinates, geographic distance (kilometers), climate distance, and georeferenced suburb ranked by \code{rank}.
}
\description{
A helper function to georeference a set of coordinates from the output of \code{seedSource} using \code{geoCode}.
}
\details{
The output data frame is first ranked in ascending order according to column name (i.e. "\code{climDist}"). The data frame is then reduced to \code{n} rows and georeferenced.
}
\note{
See \code{geoCode} for further information on the limitation of points that can be georeferenced at any one time.
}
\author{
P. A. Harrison
}
\seealso{
\code{\link{seedSource}}, \code{\link{geoCode}}
}

