% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecies2.R
\name{getSpecies2}
\alias{getSpecies2}
\title{A wrapper function to download and plot a species distribution}
\usage{
getSpecies2(species, client, plot = TRUE, region)
}
\arguments{
\item{species}{The scientific or common name of a species.}

\item{client}{The API server which to quiry to download distribution records for the selected species. See Details for available clients.}

\item{plot}{Logic whether to plot the distribution of the selected species}

\item{region}{The geographic region where the species occurs - typically the country, i.e. Australia.}
}
\value{
Returns a \code{data.frame} containing three columns: \code{species}, \code{Latitude}, \code{Longitude}.
}
\description{
A helper function for \code{speciesInPolygon}.
}
\details{
This function relies on the \code{ALA4R} and \code{rgbif} wrapper functions to quarier the API servers to download the distribution records for a specified species. The function first determines whether the species of interest occurs in Australia or outside of Australia. If the species occurs in Australia, then \code{client = "ALA"} and the Atlas of Living Australia (ALA) API server is quiried. If the species occurs outside of Australia, then \code{client = "GBIF"} and the Global Biodiversty Inforamtion Facility (GBIF) API server is quired. If \code{plot = T}, a new plotting window will open and the distirbution of the specified species will be plotted.
}
\note{
The \code{ALA4R} and \code{rgbif} packages are only available from GitHub and are not yet on CRAN. To download these packages, use Hadley Wickham's \code{devtools} package. For example:

\code{devtools::install_github("AtlasOfLivingAustralia/ALA4R")} for ALA4R

\code{devtools::install_github("ropensci/rgbif")} for rgbif
}
\section{Warning}{

The code calls \code{window()} when \code{plot = T}. On Windows operating systmes, this option is relatively stable. However, on Mac operating systmes, \code{window()} can be quite buggy and can cause R to crash.
}
\examples{
# not run

#' library(ALA4R)
library(rgbif)
library(raster)
library(sp)

# Download and plot the distribution for Eucalyptus ovata
# from the ALA server

#tmp <- PUCA::getSpecies2(species = "Eucalyptus ovata", client = "ALA",
#                         plot = TRUE, region = "Australia")

#head(tmp)
#            species Latitude Longitude
# 1 Eucalyptus ovata    -37.3     148.6
# 2 Eucalyptus ovata    -37.3     148.7
# 3 Eucalyptus ovata    -37.5     149.0
# 4 Eucalyptus ovata    -37.5     149.1
# 5 Eucalyptus ovata    -37.5     149.0
# 6 Eucalyptus ovata    -37.5     149.0

}
\author{
P. A. Harrison
}
\references{
ALA spatail portal (\url{http://spatial.ala.org.au/#})

\code{ALA4R} package (\url{https://github.com/AtlasOfLivingAustralia/ALA4R})

GBIF (\url{http://www.gbif.org/})

\code{rgbif} package (\url{https://github.com/ropensci/rgbif})
}
\seealso{
\code{\link{speciesInPolygon}}, \code{\link{plotMyRegion}}
}

