% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC_tas.R
\docType{data}
\name{LC_tas}
\alias{LC_tas}
\title{Land cover classifications for Tasmania}
\format{An object of class \code{RasterLayer} of dimension 313 x 380 x 1.}
\source{
See \url{http://modis.gsfc.nasa.gov/}
}
\usage{
LC_tas
}
\description{
A raster file containing the broad-scale landscape classifications based on the University of Maryland scheme.
}
\section{Warning}{

This data is only provided as an example.
}
\keyword{datasets}

