% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisGrids.R
\docType{data}
\name{modisGrid}
\alias{modisGrid}
\title{MODIS grid tiles}
\format{\code{SpatialPolygonsDataFrame}}
\source{
See \url{http://spatial-analyst.net/wiki/index.php?title=Main_Page}
}
\usage{
modisGrid
}
\description{
A grid of tiles overlayed on the map of the world. There are 36 horizontal (h) tiles by 17 verticle (v) tiles which describes the location of the MODIS file data. For example, the MODIS data for south-east Australia is contained in h29v12.
}
\keyword{datasets}

