\name{PUCA-package}
\alias{PUCA-package}
\alias{PUCA}
\docType{package}
\title{
\packageTitle{PUCA}
}
\description{
\packageDescription{PUCA}
}
\details{
Global climates are changing, and recent ecological restoration literature is pushing for a paradigm shift from traditional local provenancing to new restoration/reforestation provenancing strategies. The code implemented in this package employs the theory behind the climate-adjusted provenancing strategy of Prober \emph{et al.} 2015 by identifying areas of a species distribution which currently occupy analogous climates projected for the future. 

The core function to implement the climate-adjusted provenancing strategy is \code{seedSource}, which can also be implemented via the graphical user interface (GUI) by calling \code{seedSourcing_GUI}. These functions rely on the climate and distribution data being named correctly prior to been read into \code{R}, and it is highly recommended the user first read the example vignette.

\packageDESCRIPTION{PUCA}
\packageIndices{PUCA}
}
\author{
Peter A. Harrison

Maintainer: P. A. Harrison <P.A.Harrison@utas.edu.au>
}
\references{
Harrison PA, Vaillancourt RE, Harris RMB, Potts BM (in press) Predicting seed sources for future climates.\emph{Journal of Applied Ecology}.

Prober, S.M., Byrne, M., McLean, E.H., Steane, D.A., \emph{et al.} (2015) Climate-adjusted provenancing: A strategy for climate-resilient ecological restoration. \emph{Frontiers in Ecology and Evolution}, 3, 1-5.
}
