% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqLibs.R
\name{reqLibs}
\alias{reqLibs}
\title{Download/load libraries}
\usage{
reqLibs(package)
}
\arguments{
\item{package}{A list of packages to load/download}
}
\description{
\code{reqLibs} loads a list of libraries by searching the \code{R} packages directory and calling the \code{library} function. If a library is missing from the \code{R} packages directory, the function will download the library from CRAN along with all dependancies.
}
\examples{
# load the following libraries
libs <- c("ggplot2", "raster")
PUCA::reqLibs(libs)

}
\author{
P. A. Harrison
}
\seealso{
\code{\link[base]{library}}
}

