% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedSource.R
\name{seedSource}
\alias{seedSource}
\title{Find provenances using climate analoges}
\usage{
seedSource(poi, species, region, currClim, futureClim, threshold, radius,
  returnResults, verbose)
}
\arguments{
\item{poi}{A \code{data.frame} or \code{matrix} of Latitude and Longitude coordinates for the revegetation site or point of interest (POI)}

\item{species}{A list of species names that exactly match those in the global enivronement}

\item{region}{A defined region around the ecological resotration site. See Details for additional information}

\item{currClim}{A 'RasterStack' of contemporary climate layers. See Details for additional information}

\item{futureClim}{A 'RasterStack' of projected future climate layers. See Details for additional information}

\item{threshold}{A numeric value to transform the climate distance into a binary value}

\item{radius}{The radius of the buffer that is drawn around the revegetation site using \code{drawBuffer}}

\item{returnResults}{The number of top ranking distribution points to print to the console based on the climate distance between the revegetaion site and each distribution point}

\item{verbose}{Logic whether to print a log to the screen}
}
\value{
There are numerious objects that are assigned to the \code{.GlobalEnv}, but the most important are:

1. The Principal Components Analysis (PCA) for the region, species, and revegetation site. The PCA is assigned for each species and begins with '\code{PCA_}'. Calling \code{ls(pattern = "PCA_+", envir = .GlobalEnv)} will list the various PCA objects.

2. The output of the \code{seedSource} function. The output is uniquely assigned using the follow convention: '\code{output_species}', species name, whether its for the contemporary (\code{revegCurrent}) or future (\code{revegFuture}). For the future predictions there will be an additional two prefixs specifying the global circulation model (or multi-model mean (MMM) used and the time slice. Calling \code{ls(pattern = "output_+", envir = .GlobalEnv)} will list the various output objects.
}
\description{
The core function of the \code{PUCA} package to implement the climate-adjusted provenancing strategy of Prober \emph{et al.} 2015.
}
\details{
The function first creates a uniform 'SpatialPoints' grid across the defined \code{region}, matching the resolution of the climate layers (i.e. if \code{res(currClim)[1]} equals 0.01 then the grid points are 0.01 apart). The contemporary climate is then extracted and passed to the  Principal Components Analysis (PCA), which is implemented using the \code{PCA} function of package \code{FactoMineR}. It is this PC climate space that is used to define the Euclidean distance between the revegetation site and each species distribution point. \strong{Thus, the definition of \code{region} is important and requires due consideration}. A region may de defined by (i) a specific distance from the revegetation site (see \code{drawRegion}), (ii) using the eco-regions of the world (i.e. http://wwf.panda.org/about_our_earth/ecoregions/ecoregion_list/), (iii) or bio-regions of a country (i.e. Australias IBRA bioregions -https://www.environment.gov.au/land/nrs/science/ibra/australias-bioregions-maps).

The RasterStacks contain the contemporary and future climate projections. The only requirement for these RasterStacks is the number of climate variables for the projected future climate RasterStack must match the same number of climate variables in the contemporary climate RasterStack. For example, if the contemporary climate RasterStack contained 19 climate varaibles, then the future climate RasterStack \strong{must} contain projected values for the same 19 varaibles.

By default, if more than one global circulation model (i.e. future projection) is provided to the function, it will automatically create an ensemble of these models. The ensembled layers will thus be a multi-model mean by averaging each cell projection.
}
\examples{
# See vignette for a detailed example

}
\author{
P. A. Harrison, B. M. Potts
}
\references{
Harrison PA, Vaillancourt RE, Harris RMB, Potts BM (in press) Predicting seed sources for future climates.\emph{Journal of Applied Ecology}.

Prober SM, Byrne M, McLean EH, Steane DA, \emph{et al.} (2015) Climate-adjusted provenancing: A strategy for climate-resilient ecological restoration. \emph{Frontiers in Ecology and Evolution}, 3, 1-5.
}
\seealso{
\code{\link{drawRegion}}, \code{\link{drawBuffer}}, \code{\link[FactoMineR]{PCA}}, \code{\link{seedSource_GUI}}
}

