% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_seedSourcing.R
\name{seedSource_GUI}
\alias{seedSource_GUI}
\title{Graphical user interface for \code{seedSource}}
\usage{
seedSource_GUI()
}
\description{
A graphical user interface (GUI) implementation of the \code{seedSource} function.
}
\details{
See the vignette for details on how to use the GUI. Additonally, see \code{seedSource} for futher details.
}
\examples{
# Run the GUI
#PUCA::seedSource_GUI()

}
\author{
P. A. Harrison
}
\seealso{
\code{\link{seedSource}}
}

