% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNames.R
\name{seqNames}
\alias{seqNames}
\title{Create a sequence of names}
\usage{
seqNames(text, quantity)
}
\arguments{
\item{text}{A text variable that is repeated \code{quantity} times}

\item{quantity}{The maximum numerical value for the sequence}
}
\description{
\code{seqNames} creates a sequence of names that has length \code{quantity}
}
\details{
This code was adpated from Zurr et al (2013).
}
\examples{
PUCA::seqNames('bio', 19)

# Returns
# "bio1"  "bio2"  "bio3"  "bio4"  "bio5"  "bio6"  "bio7"  "bio8"  "bio9"
# "bio10" "bio11" "bio12" "bio13" "bio14" "bio15" "bio16" "bio17" 
# "bio18" "bio19"

}
\author{
P. A. Harrison
}
\references{
Zurr AF, Hilbe JM, Leno EA (2013) A beginner's guide to GLM and GLMM with R. Highland Statistics Ltd (Newburgh, UK).
}

