% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesInPolygon.R
\name{speciesInPolygon}
\alias{speciesInPolygon}
\title{Find what species occur within a polygon}
\usage{
speciesInPolygon(x, y, r, numVert = 1000)
}
\arguments{
\item{x}{A longitude coordinate}

\item{y}{A latitude coordinate}

\item{r}{The radius (in kilometers) of the polgon}

\item{numVert}{The number of vertices to create the polygon}
}
\value{
A \code{data.frame} containing the kingdom the species belongs to, the species name, their common name (if known), and the number of occurnance records
}
\description{
A wrapper function that calls \code{ALA4R} to searche the Atlas of Living Australia for species occurnances within a polygon of size \code{radius}.
}
\details{
This feature is currently only available for searches within Australia. Later versions will implement a call to the GBIF API server. Use \code{getSpecies2} to download occurance records for a species outside of Australia. 

This function relies on the wapper function \code{specieslist} of package \code{ALA4R}, and we greatly acknowledge the creator(s) of this package.
}
\examples{
# Not run
#library(sp)
#library(raster)

#tmp <- PUCA::speciesInPolygon(x = 147.4686, y = -42.008, r = 5)
#head(tmp)

}
\author{
P. A. Harrison
}

