% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMypath.R
\name{splitMyDir}
\alias{splitMyDir}
\title{Split the path of a directory}
\usage{
splitMyDir(Dir)
}
\arguments{
\item{Dir}{The path name of a directory}
}
\value{
A list of folder names, with the lowest nested folder occuring first,
and returns the reverse order of the directory such that the lowest level is first.
}
\description{
A helper function that splits the path of a direcory based on either the '\\\' or '/'.
}
\examples{
# Example using '\\'
tmp <- "C:\\\\Test\\\\Test1\\\\test2\\\\test3"
PUCA::splitMyDir(Dir = tmp)
# [1] "test3" "test2" "Test1" "Test"  "C:"   

# Example using '/'
tmp <- "C:/Test/Test1/test2/test3"
PUCA::splitMyDir(Dir = tmp)
# [1] "test3" "test2" "Test1" "Test"  "C:" 

}
\author{
P. A. Harrison
}

